requirements.txt
pytest
selenium
openpyxl
pandas
python-dotenv
